Remove payment method cell from Checkout
