+++ uisdk/src/main/res/layout/uisdk_booking_checkout_view.xml
@@ -58,13 +58,6 @@
                     android:id="@+id/bookingCheckoutPassengerView"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
-                    android:layout_marginEnd="@dimen/kh_uisdk_spacing_small"
-                    android:layout_weight="1" />
-
-                <com.karhoo.uisdk.screen.booking.checkout.payment.BookingPaymentView
-                    android:id="@+id/bookingRequestPaymentDetailsWidget"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
                     android:layout_weight="1" />
             </LinearLayout>
